+++ module/data/actor-character.mjs
@@ -207,6 +207,7 @@ export default class SdmCharacter extends SdmActorBase {
         min: 0
       }),
       max: new fields.NumberField({ ...requiredInteger, initial: 0, min: 0 }),
+      max_limit: new fields.NumberField({ ...requiredInteger, initial: 0, min: 0 }),
       bonus: new fields.NumberField({ ...requiredInteger, initial: 0 }),
       imbued: new fields.NumberField({ ...requiredInteger, initial: 0, min: 0 }),
     });